# packages/relayer/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Taiko Bridge Relayer API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "https://community.taiko.xyz/",
      "email": "info@taiko.xyz"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/taikoxyz/taiko-mono/blob/main/LICENSE"
    },
    "version": "1.0"
  },
  "host": "relayer.hekla.taiko.xyz",
  "paths": {
    "/blockInfo": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Get block info",
        "operationId": "get-block-info",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/http.getBlockInfoResponse"
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "summary": "Get events by address",
        "operationId": "get-events-by-address",
        "parameters": [
          {
            "type": "string",
            "description": "address to query",
            "name": "address",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "msgHash to query",
            "name": "msgHash",
            "in": "query"
          },
          {
            "type": "string",
            "description": "chainID to query",
            "name": "chainID",
            "in": "query"
          },
          {
            "type": "string",
            "description": "eventType to query",
            "name": "eventType",
            "in": "query"
          },
          {
            "type": "string",
            "description": "event to query",
            "name": "event",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/paginate.Page"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "http.blockInfo": {
      "type": "object",
      "properties": {
        "chainID": {
          "type": "integer"
        },
        "latestBlock": {
          "type": "integer"
        },
        "latestProcessedBlock": {
          "type": "integer"
        }
      }
    },
    "http.getBlockInfoResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/http.blockInfo"
          }
        }
      }
    },
    "paginate.Page": {
      "type": "object",
      "properties": {
        "error": {
          "type": "boolean"
        },
        "error_message": {
          "type": "string"
        },
        "first": {
          "type": "boolean"
        },
        "items": {},
        "last": {
          "type": "boolean"
        },
        "max_page": {
          "type": "integer"
        },
        "page": {
          "type": "integer"
        },
        "size": {
          "type": "integer"
        },
        "total": {
          "type": "integer"
        },
        "total_pages": {
          "type": "integer"
        },
        "visible": {
          "type": "integer"
        }
      }
    }
  }
}